Fixing schmidt decomposition function.
