+++ toqito/state_ops/__init__.py
@@ -1,2 +1,3 @@
 """Operations on quantum states."""
 from toqito.state_ops.pure_to_mixed import pure_to_mixed
+from toqito.state_ops.schmidt_decomposition import schmidt_decomposition
