+++ toqito/state_props/__init__.py
@@ -10,7 +10,6 @@ from toqito.state_props.concurrence import concurrence
 from toqito.state_props.negativity import negativity
 from toqito.state_props.log_negativity import log_negativity
 from toqito.state_props.purity import purity
-from toqito.state_props.schmidt_decomposition import schmidt_decomposition
 from toqito.state_props.schmidt_rank import schmidt_rank
 from toqito.state_props.von_neumann_entropy import von_neumann_entropy
 from toqito.state_props.entanglement_of_formation import entanglement_of_formation
